feat(api): continue the split
- extract electrum.cfg.cpp/.hpp from coins.cfg.cpp/.hpp to simplify electrum configuration
- Add rpc.electrum.cpp/.hpp that represents the mm2 electrum rpc call
- Remove electrum rpc from mm2.cpp/.hpp
- Add the sources to the CMakeLists.txt
