+++ src/atomicdex/api/mm2/mm2.cpp
@@ -159,32 +159,6 @@ namespace
     }
 } // namespace
 
-//! Implementation RPC [electrum]
-namespace mm2::api
-{
-    //! Serialization
-    void
-    to_json(nlohmann::json& j, const electrum_request& cfg)
-    {
-        j["coin"]       = cfg.coin_name;
-        j["servers"]    = cfg.servers;
-        j["tx_history"] = cfg.with_tx_history;
-        if (cfg.coin_type == CoinType::QRC20)
-        {
-            j["swap_contract_address"] = cfg.is_testnet ? cfg.testnet_qrc_swap_contract_address : cfg.mainnet_qrc_swap_contract_address;
-        }
-    }
-
-    //! Deserialization
-    void
-    from_json(const nlohmann::json& j, electrum_answer& cfg)
-    {
-        j.at("address").get_to(cfg.address);
-        j.at("balance").get_to(cfg.balance);
-        j.at("result").get_to(cfg.result);
-    }
-} // namespace mm2::api
-
 //! Implementation RPC [disable_coin]
 namespace mm2::api
 {
